moved blocs upper level
