+++ lib/ui/blocs/export.dart
@@ -3,6 +3,7 @@ export 'home/export.dart';
 export 'mnemonic/export.dart';
 export 'navigator/export.dart';
 export 'notifications/export.dart';
-export 'posts/export.dart';
 export 'recover_account/export.dart';
+export 'posts_list/export.dart';
+export 'post_details/export.dart';
 export 'simple_bloc_delegate.dart';
