+++ lib/ui/blocs/post_details/post_details_state.dart
