--- lib/ui/blocs/posts/export.dart
@@ -1,2 +1,0 @@
-export 'post_details/bloc.dart';
-export 'posts_list/bloc.dart';
